+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/ContactsControllerTests.cs
@@ -1,5 +1,4 @@
-﻿using System.Security.Claims;
-using AutoFixture;
+﻿using AutoFixture;
 using Dfe.Complete.Api.Tests.Integration.Customizations;
 using Dfe.Complete.Client.Contracts;
 using Dfe.Complete.Infrastructure.Database;
@@ -8,6 +7,7 @@ using Dfe.Complete.Tests.Common.Customizations.Models;
 using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Attributes;
 using GovUK.Dfe.CoreLibs.Testing.Mocks.WebApplicationFactory;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 using ContactId = Dfe.Complete.Domain.ValueObjects.ContactId;
 
 namespace Dfe.Complete.Api.Tests.Integration.Controllers;